requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
transformers==4.36.2
accelerate==0.26.1
requests==2.31.0
jinja2==3.1.3
python-multipart==0.0.6
soundfile==0.12.1
librosa==0.10.1
numpy==1.26.3
aiohttp==3.9.1
ollama==0.1.0
coqui_tts==0.17.0
